Revert "Move server init into influx pkg"
